Fix errata involving early definition of NewCLI
Previously, `func NewCLI` was moved to be defined early but downstream
references caused inconsistencies in the text. Theses changes fix the
inconsistencies by removing the early definition and fixing the code
references in the text so that the code is sensible and working for the
reader writing alongside the text.