fix CI for forks
As an upstream, this repo depends on itself, but for a fork, it depends
on this repo. In Travis CI, `install: true` skipps `go get` for Go
projects, so the forks won't get the dependence and will cause the
CI failure. This is fixed by installing dependence in `before_install`.

Ref: https://docs.travis-ci.com/user/job-lifecycle/#skipping-the-installation-phase