Add missing allRomanNumerals data in section
This commit addresses #355 in which @twirrim notes that
`allRomanNumerals` wasn't defined in
<https://quii.gitbook.io/learn-go-with-tests/go-fundamentals/roman-numerals#write-enough-code-to-make-it-pass-7>
and was shown further down making it a bit confusing for readers. The
missing `allRomanNumerals` has now been added to ensure that readers
have the context they may need to show how `RomanNumerals` methods can
be called on `allRomanNumerals` since it is now of type `RomanNumerals`.

Fixes #355