Add clarifying image above secondHandPoint in math
The unit circle picture that is provided in
<https://quii.gitbook.io/learn-go-with-tests/go-fundamentals/math#write-enough-code-to-make-it-pass-3>
is the standard unit circle which measures angles from the X axis
instead of the Y axis. This may cause readers to use cosine when
calculating X and sin when calculating Y instead of the other way
around. This change adds the image in which the triangle is flipped and
the angle is measured from the Y axis (12 o'clock) and in which X is
clearly shown as sin(α) and Y is shown as cos(α)