Added additional context to example changes for the 'it returns acepted on POST'
As written it was a bit confusing the exact change that needed to be made, so I dug into the example code to look up where it was to go. The main bit of confusion was the name of the test changed. 

I've added the rest of the function around the previous code snippet to add additional context