Add quotes to switch cases
I have tried `switch` with and without quotes, and the compiler complains without quotes. The docs on `switch` also has quotes.

(As an aside, should I open an issue first? I am using the Edit button on GH, but I am happy to open an issue in the future if that is the preferred way to do things. As a meta-aside, should I open an issue proposing a guideline for editing doc?)