fix: Consistent example code for SumAll() example
The first code example showing the shape of the `SumAll()` function has a named return value in it of `(sum []int)` where later examples drop the named return value. If learners are not careful and notice the change, they may overlook the different return signature and get surprised by a `NoNewVar` error later on in the process when `sum := ...` is written.
https://pkg.go.dev/golang.org/x/tools/internal/typesinternal#NoNewVar

This changes the initial entry to line up with the rest of the examples in the document.