fix(math): clarify which test file to use
A section of the math chapter says to add an Svg struct to a
`clockface_assembly_test` file, but that's not the name of a test file
that's used in the clockface example. Because the next state of the code
(v7b) shows that struct (and the corresponding code) in the
`clockface_acceptance_test` file, this MR updates the name of the
mentioned test file from `clockface_assembly_test` to
`clockface_acceptance_test`.