Correct syntax error in select.md (#745)
In the previous section, an error return was added to the `Racer()` function. In the text underneath this code block, there is a note that says. 

"Note that we've also handled the error return in our original text, we're using `_` for now to ensure the tests will run.

I think the error was not actually ignored with an underscore as you intended.