Add note on potential go benchmark optimisations  (#781)
* Fix style of note on benchmarks running in sequence

Note was italicized and didn't have a colon. This didn't match the other case of a note
in the itegers chaper which was.

* Add a note on go optimizing benchmarks

Added a note on go optimizing benchmarks and a link to Dave Cheney's blog post
on how to write benchmarks for how to make sure it doesn't. This optimization seems
rare from reading issues online but does happen so warning would be good to let people know.

* Fix wording to give a better sense of what go can do to optimize

* Fix missing period